+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -1,2 +1,8 @@
 CHANGELOG
 =========
+
+3.3.0
+-----
+
+ * Deprecated class name support in `WorkflowRegistry::add()` as second parameter.
+   Wrap the class name in an instance of ClassInstanceSupportStrategy instead.